Just refactoring the loggedIn stuff into its own package
