+++ src/main/java/app/AppBuilder.java
@@ -14,9 +14,9 @@ import interface_adapter.change_password.BackToLoggedInController;
 import interface_adapter.change_password.ChangePasswordController;
 import interface_adapter.change_password.ChangePasswordPresenter;
 import interface_adapter.change_password.ChangePasswordViewModel;
-import interface_adapter.change_password.LoggedInPresenter;
-import interface_adapter.change_password.LoggedInViewModel;
-import interface_adapter.change_password.ToPasswordSettingsController;
+import interface_adapter.logged_in.LoggedInPresenter;
+import interface_adapter.logged_in.LoggedInViewModel;
+import interface_adapter.logged_in.ToPasswordSettingsController;
 import interface_adapter.home_view.GoToLoginController;
 import interface_adapter.home_view.GoToSignUpController;
 import interface_adapter.home_view.HomeViewModel;
