+++ src/main/java/interface_adapter/change_password/ChangePasswordPresenter.java
@@ -1,7 +1,7 @@
 package interface_adapter.change_password;
 
 import interface_adapter.ViewManagerModel;
-import interface_adapter.signup.SignupState;
+import interface_adapter.logged_in.LoggedInViewModel;
 import use_case.change_password.ChangePasswordOutputBoundary;
 import use_case.change_password.ChangePasswordOutputData;
